+++ pages/api/team-assignments/create.ts
@@ -6,6 +6,7 @@ import {
   RingWithAssignments,
   UmpirePagePlayer
 } from "../../../types/umpirepage";
+import { umpirePagePlayerSelect } from "../../../lib/prisma-selects";
 
 const isCurrentUserAuthorized = async (tournamentId, req, res) => {
   const session = await getServerSession(req, res, authConfig);
@@ -87,38 +88,7 @@ export default async function handler(
   });
 
   const updatedPlayers: UmpirePagePlayer[] = await prisma.player.findMany({
-    select: {
-      id: true,
-      title: true,
-      alias: true,
-      state: true,
-      user: {
-        select: {
-          id: true,
-          firstName: true,
-          lastName: true
-        }
-      },
-      team: {
-        select: {
-          id: true,
-          name: true
-        }
-      },
-      umpire: {
-        select: {
-          id: true,
-          user: {
-            select: {
-              id: true,
-              firstName: true,
-              lastName: true
-            }
-          }
-        }
-      },
-      targets: true
-    }
+    select: umpirePagePlayerSelect
   });
 
   const playerRings: RingWithAssignments[] =
