+++ pages/api/team-rings/create.ts
@@ -7,6 +7,7 @@ import {
   TeamRingWithAssignments,
   UmpirePagePlayer
 } from "../../../types/umpirepage";
+import { umpirePagePlayerSelect } from "../../../lib/prisma-selects";
 
 const isCurrentUserAuthorized = async (tournamentId, req, res) => {
   const session = await getServerSession(req, res, authConfig);
@@ -89,38 +90,7 @@ export default async function handler(
     });
 
     const updatedPlayers: UmpirePagePlayer[] = await prisma.player.findMany({
-      select: {
-        id: true,
-        title: true,
-        alias: true,
-        state: true,
-        user: {
-          select: {
-            id: true,
-            firstName: true,
-            lastName: true
-          }
-        },
-        team: {
-          select: {
-            id: true,
-            name: true
-          }
-        },
-        umpire: {
-          select: {
-            id: true,
-            user: {
-              select: {
-                id: true,
-                firstName: true,
-                lastName: true
-              }
-            }
-          }
-        },
-        targets: true
-      }
+      select: umpirePagePlayerSelect
     });
 
     const playerRings: RingWithAssignments[] =
