+++ pages/api/tournament/assignments.ts
@@ -3,6 +3,7 @@ import type { NextApiRequest, NextApiResponse } from "next";
 import { getServerSession } from "next-auth";
 import { authConfig } from "../auth/[...nextauth]";
 import { UmpirePagePlayer } from "../../../types/umpirepage";
+import { umpirePagePlayerSelect } from "../../../lib/prisma-selects";
 
 const isCurrentUserAuthorized = async (tournamentId, req, res) => {
   const session = await getServerSession(req, res, authConfig);
@@ -86,38 +87,7 @@ export default async function handler(
     });
 
     const updatedPlayers: UmpirePagePlayer[] = await prisma.player.findMany({
-      select: {
-        id: true,
-        title: true,
-        alias: true,
-        state: true,
-        user: {
-          select: {
-            id: true,
-            firstName: true,
-            lastName: true
-          }
-        },
-        team: {
-          select: {
-            id: true,
-            name: true
-          }
-        },
-        umpire: {
-          select: {
-            id: true,
-            user: {
-              select: {
-                id: true,
-                firstName: true,
-                lastName: true
-              }
-            }
-          }
-        },
-        targets: true
-      }
+      select: umpirePagePlayerSelect
     });
 
     res.json({ savedAssignments, players: updatedPlayers });
